Fix a critical bug where backing out during the routine causes it to hang indefinitely
Reason: Used sendBlocking() during development, to send events to the ui layer.
When backing out though, the activity is destroyed and the channel consumer is no
longer there.  Since the channel is constructed with default parameters, the buffer
strategy is RENDEZVOUS, or rather, no strategy, and the block hangs forever, during
dispatchFrame. Since that is on the main, thread, it will never come back

Fix: launch and send in a coroutine.  I don't think it's the absolute cleanest fix
but I need to redo the timing infrastructure anyways, and cleaning up what executes
where and when